requirements.txt
--extra-index-url https://admin:..MM!aws!01..@python.marketmuse.com/pypi

clickhouse-connect
aiochclient[aiohttp]
clickhouse-driver[numpy]
faker
gdown

historical_serps_data
SQLAlchemy==1.4.46
serp_vault==5.0.27
python-dotenv
tldextract

asyncpg==0.28.0

pandas==2.0.3
s3fs==2023.6.0
sentry-sdk==1.28.1